+++ src/lib/environment/index.ts
@@ -21,9 +21,9 @@
 */
 
 import { LogLevel } from '@lib/logger';
-import dotenvLoader from '@lib/utility/dotenvLoader';
-import typeConverters from '@lib/utility/typeConverter';
 import { projectDirectoryName } from '@lib/directories';
+import dotenvLoader from '@lib/environment/dotenvLoader';
+import typeConverters from '@lib/environment/typeConverter';
 
 import * as fs from 'fs';
 import * as path from 'path';
